Add Base classifier (#69)
* Added base class

* Added base class
