+++ autokeras/__init__.py
@@ -1,1 +1,1 @@
-from autokeras.classifier import ImageClassifier
+from autokeras.image_classifier import ImageClassifier
