+++ examples/mnist.py
@@ -1,5 +1,5 @@
 from keras.datasets import mnist
-from autokeras.classifier import ImageClassifier
+from autokeras.image_classifier import ImageClassifier
 
 if __name__ == '__main__':
     (x_train, y_train), (x_test, y_test) = mnist.load_data()
